refactor(runtime): CliError を error.rs モジュールに分離
main.rs はエントリポイントと CLI 定義に集中させ、エラー型は責務単位でファイル分割。今後パイプライン・ドライバー起動・IPC などエラー種別が増えていく見込みのため、早めに置き場所を確立しておく。
