+++ crates/midori-runtime/src/main.rs
@@ -1,7 +1,11 @@
+mod error;
+
 use std::path::{Path, PathBuf};
 use std::process::ExitCode;
 
 use clap::{Parser, Subcommand, ValueEnum};
+
+use crate::error::CliError;
 
 #[derive(Parser, Debug)]
 #[command(
@@ -77,36 +81,6 @@ fn run(profile_path: &Path) -> Result<(), CliError> {
             source,
         })?;
     Ok(())
-}
-
-#[derive(Debug)]
-enum CliError {
-    ReadProfile {
-        path: PathBuf,
-        source: std::io::Error,
-    },
-}
-
-impl std::fmt::Display for CliError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::ReadProfile { path, source } => {
-                write!(
-                    f,
-                    "プロファイルの読み込みに失敗しました ({}): {source}",
-                    path.display()
-                )
-            }
-        }
-    }
-}
-
-impl std::error::Error for CliError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        match self {
-            Self::ReadProfile { source, .. } => Some(source),
-        }
-    }
 }
 
 #[cfg(test)]
